+++ examples/Rules-method-stopTrialCITDsamples.R
@@ -1,3 +1,5 @@
+# nolint start
+
 ##define the stopping rules based on the 'StoppingTDCIRatio' class
 ##Using only DLE responses with samples
 ## we need a data object with doses >= 1:
@@ -26,6 +28,7 @@ RecommendDose<-nextBest(tdNextBest,doselimit=max(data@doseGrid),samples=samples,
 ##use 'stopTrial' to determine if the rule has been fulfilled
 ##use 0.3 as the target proability of DLE at the end of the trial
 
-stopTrial(stopping=myStopping,dose=RecommendDose$nextdose,
+stopTrial(stopping=myStopping,dose=RecommendDose$next_dose_drt,
           samples=samples,model=model,data=data)
-## RecommendDose$nextdose refers to the next dose obtained in RecommendDose
+
+# nolint end